requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
langchain-openai
langchain
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
prometheus-client
redis
structlog
pydantic
pytest
pytest-mock
